Add session management with redis
Добавлено хранение сессий и авторизация по ним

+++ Backend/src/main/java/com/messenger/prism/repository/MessageRepo.java
@@ -3,5 +3,5 @@ package com.messenger.prism.repository;
 import com.messenger.prism.entity.MessageEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface MessageRepo extends CrudRepository<MessageEntity, Long> {
+public interface MessageRepo extends CrudRepository<MessageEntity, Integer > {
 }
